fix bug 1432 and rename files to use proper convention
