+++ transforms/universal/tokenization/Makefile
@@ -17,5 +17,13 @@ TRANSFORM_NAME=$(shell basename `pwd`)
 run-cli-sample: 
 	make venv
 	source venv/bin/activate && \
-	$(PYTHON) -m dpk_$(TRANSFORM_NAME).transform_python \
+	$(PYTHON) -m dpk_$(TRANSFORM_NAME).runtime \
             --data_local_config "{ 'input_folder' : 'test-data/ds01/input', 'output_folder' : 'output'}"
+
+run-list-sample: 
+	make venv
+	source venv/bin/activate && \
+	$(PYTHON) -m dpk_$(TRANSFORM_NAME).runtime \
+			--tkn_doc_content_column  curr_file_contents \
+			--tkn_doc_id_column pulls_and_issues_id \
+            --data_local_config "{ 'input_folder' : 'test-data/list/input', 'output_folder' : 'output'}"
