+++ transforms/universal/tokenization/dpk_tokenization/__init__.py
@@ -1,3 +1,3 @@
 # SPDX-License-Identifier: Apache-2.0
 from .transform import *
-from .transform_python import *
+from .runtime import *
