+++ transforms/universal/tokenization/dpk_tokenization/ray/__init__.py
@@ -1,2 +1,2 @@
 # SPDX-License-Identifier: Apache-2.0
-from .transform import *
+from .runtime import *
