+++ transforms/universal/tokenization/kfp_ray/tokenization_wf.py
@@ -27,7 +27,7 @@ from workflow_support.compile_utils import (
 S3_SECRET = "s3-secret"
 
 # the name of the job script
-EXEC_SCRIPT_NAME: str = "-m dpk_tokenization.ray.transform"
+EXEC_SCRIPT_NAME: str = "-m dpk_tokenization.ray.runtime"
 
 task_image = "quay.io/dataprep1/data-prep-kit/tokenization-ray:latest"
