+++ transforms/universal/tokenization/test-data/list/expected/metadata.json
@@ -1,0 +1,63 @@
+{
+  "pipeline": "pipeline_id",
+  "job details": {
+    "job category": "preprocessing",
+    "job name": "Tokenization",
+    "job type": "pure python",
+    "job id": "job_id",
+    "start_time": "2025-08-26 12:09:40",
+    "end_time": "2025-08-26 12:09:41",
+    "status": "success"
+  },
+  "code": {
+    "github": "UNDEFINED",
+    "build-date": "UNDEFINED",
+    "commit_hash": "UNDEFINED",
+    "path": "UNDEFINED"
+  },
+  "job_input_params": {
+    "tokenizer": "hf-internal-testing/llama-tokenizer",
+    "tokenizer_args": null,
+    "doc_id_column": "pulls_and_issues_id",
+    "doc_content_column": "curr_file_contents",
+    "text_lang": "en",
+    "chunk_size": 0,
+    "checkpointing": false,
+    "max_files": -1,
+    "random_samples": -1,
+    "files_to_use": [
+      ".parquet"
+    ],
+    "num_processors": 0
+  },
+  "execution_stats": {
+    "cpus": 29.6,
+    "gpus": 0,
+    "memory": 13.6,
+    "object_store": 0,
+    "execution time, min": 0.015
+  },
+  "job_output_stats": {
+    "source_files": 1,
+    "source_size": 316400,
+    "result_files": 1,
+    "result_size": 108120,
+    "processing_time": 0.205,
+    "num_files": 1,
+    "num_rows": 1,
+    "num_tokenized_rows": 1,
+    "num_empty_rows": 0,
+    "num_tokens": 134921,
+    "num_chars": 250171,
+    "source_doc_count": 1,
+    "result_doc_count": 1
+  },
+  "source": {
+    "name": "/Users/touma/data-prep-kit-pkg/transforms/universal/tokenization/test-data/list/input",
+    "type": "path"
+  },
+  "target": {
+    "name": "/Users/touma/data-prep-kit-pkg/transforms/universal/tokenization/output",
+    "type": "path"
+  }
+}
