+++ transforms/universal/tokenization2arrow/requirements.txt
@@ -1,3 +1,3 @@
 transformers>=4.38.2
-data-prep-toolkit-transforms[tokenization]
+data-prep-toolkit-transforms[tokenization]>=1.1.4.dev0
 python-dotenv
